feat: kind classes renamed
