+++ packages/core/src/vcdm/encoding/rlp/kind/ScalarKind.ts
