+++ packages/core/src/vcdm/encoding/rlp/kind/hexblob/index.ts
@@ -1,0 +1,4 @@
+export * from './HexBlobKind';
+export * from './FixedHexBlobKind';
+export * from './OptionalFixedHexBlobKind';
+export * from './CompactFixedHexBlobKind';
