--- packages/core/src/vcdm/encoding/rlp/kind/hexblobkinds/index.ts
@@ -1,4 +1,0 @@
-export * from './hexblob';
-export * from './fixedhexblob';
-export * from './optionalfixedhexblob';
-export * from './compactfixedhexblob';
